# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  goals.atualizarProgressoRequest:
    properties:
      current:
        type: number
    type: object
  goals.criarMetaRequest:
    properties:
      category:
        type: string
      color:
        type: string
      current:
        type: number
      deadline:
        type: string
      description:
        type: string
      month:
        type: integer
      scope:
        type: string
      target:
        type: number
      title:
        type: string
      unit:
        type: string
      year:
        type: integer
    type: object
  goals.editarMetaRequest:
    properties:
      category:
        type: string
      color:
        type: string
      current:
        type: number
      deadline:
        type: string
      description:
        type: string
      target:
        type: number
      title:
        type: string
      unit:
        type: string
    type: object
  habits.criarHabitoRequest:
    properties:
      color:
        type: string
      iconKey:
        type: string
      name:
        type: string
    type: object
  habits.editarHabitoRequest:
    properties:
      color:
        type: string
      iconKey:
        type: string
      name:
        type: string
    type: object
  habits.toggleRequest:
    properties:
      date:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: 'API do BeePlanner — autenticação via JWT (Bearer). Todas as rotas
    /api/* (exceto /auth) exigem o header Authorization: Bearer <token>.'
  title: BeePlanner API
  version: "1.0"
paths:
...
  /goals:
    get:
      parameters:
      - description: annual ou monthly
        in: query
        name: scope
        type: string
      - description: Ano
        in: query
        name: year
        type: integer
      - description: Mês (1-12)
        in: query
        name: month
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar metas
      tags:
      - Goals
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados da meta
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goals.criarMetaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar meta
      tags:
      - Goals
  /goals/{id}:
    delete:
      parameters:
      - description: ID da meta
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Deletar meta
      tags:
      - Goals
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID da meta
        in: path
        name: id
        required: true
        type: string
      - description: Valor atual
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goals.atualizarProgressoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar progresso da meta
      tags:
      - Goals
    put:
      consumes:
      - application/json
      parameters:
      - description: ID da meta
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/goals.editarMetaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Editar meta
      tags:
      - Goals
  /habits:
    get:
      description: Retorna todos os hábitos com conclusões dos últimos 60 dias e sequência
        atual
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar hábitos
      tags:
      - Habits
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do hábito
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/habits.criarHabitoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar hábito
      tags:
      - Habits
  /habits/{id}:
    delete:
      parameters:
      - description: ID do hábito
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deletar hábito
      tags:
      - Habits
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do hábito
        in: path
        name: id
        required: true
        type: string
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/habits.editarHabitoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Editar hábito
      tags:
      - Habits
  /habits/{id}/toggle:
    post:
      consumes:
      - application/json
      description: Marca ou desmarca o hábito numa data específica (YYYY-MM-DD)
      parameters:
      - description: ID do hábito
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/habits.toggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Toggle conclusão do hábito
      tags:
      - Habits
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <seu_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"